refactor: fix 'dead code'
