+++ potter/src/App/store/api/api.tsx
@@ -1,5 +1,4 @@
 import {
-  FetchBaseQueryMeta,
   createApi,
   fetchBaseQuery,
 } from "@reduxjs/toolkit/query/react"
