+++ potter/src/pages/Header/Header.tsx
@@ -10,17 +10,12 @@ import {
 } from "../../utils/context/CharacterContext"
 import { useDispatch, useSelector } from "react-redux"
 import { RootState } from "../../App/store/store"
-import {
-  checkFavorite,
-  checkHistory,
-  checkUserActive,
-  loadUserData,
-} from "../../App/store/userSlice"
+import { checkHistory, checkUserActive } from "../../App/store/userSlice"
 import {
   addHistory,
   getUserActive,
   removeUser,
-} from "../../utils/LS/forWorkWithUser"
+} from "../../utils/workUser/forWorkWithUser"
 import useDebounce from "./useDebounce"
 
 export const Header = () => {
